version 3.25.0 This update breaks some things: - Stage is now saved to the database so that it sticks across server restarts (WARNING: You will need to delete data.mv.db, doing so will delete all placed generators in your world)
- Sugar canes can now only be placed near water (Like in vanilla)

- Cactus can only be placed on sand (Like in vanilla)

- Sugar cane and cactus are now subject to physics.

- Crops can only be placed on tilted soil. (Like in vanilla)

- Cocoa can only be placed on jungle wood/logs (Like in vanilla)

- Generators are now removed if broken by water/lava, before fluid flow was interrupted.

Fixed bamboo not behaving properly

Fixed 'idle' option and renamed it to 'regrow' (if regrow is false when mined the block will not regenerate and you will have to do so manually using setStage, nextStage or previousStage in on-break trigger)

This update might seem like a downgrade but it is necessary to make a more solid and less buggy system. I hope to compensate with more interesting features in the future.